dlcservice: Handle Preallocated size (DLC_PREALLOC_BLOCKS) increase scenario 1
This is the simplest of the cases to handle when a DLC package
increases its already set DLC_PREALLOC_BLOCKS for various reasons.

This scenario builds on top of the assumption that update_engine has no
problem when applying an update that increases the inactive DLC image.
(This scenario will be handled in the future)

Example flow:
 DLC_A size = DLC_B size = 5MB
 Update applied to inactive image DLC_B, new DLC_B size = 10MB
 On reboot DLC_B is now active, but DLC_A size + 5MB = DLC_B size.
 dlcservice should increase DLC_A to 10MB and catch this issue as early
 as possible before another update is applied where update_engine would
 write to DLC_A with potential failure.

BUG=chromium:943780
TEST=# unittest + tast

Change-Id: I67617197722e56fd55b5ecbd6990282a0cfa307c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2006024
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>