shill: device_info: Unify Device destruction for physical technologies
Prior to this change, Cellular followed a separate destruction path
from all the other physical Devices. Essentially, Modem would control
the ownership of the Cellular instance. During Modem destruction, it
would call DeviceInfo::DeregisterDevice to cause DeviceInfo to
relinquish ownership of the Cellular instance and perform what would
normally be done in DeviceInfo::RemoveInfo. Destruction of the
Cellular instance itself (and hence the logic in ~Cellular) would be
performed along with Modem destruction. Then later, an RTNL link
delete message would finish the job of fully removing the
DeviceInfo::Info corresponding to the Cellular interface (presumably
this was done so that DeviceInfo wouldn't spuriously think it got a
link delete message for an unknown interface).

This change causes all physical Devices to be destructed by
DeviceInfo::RemoveInfo triggered by an RTNL link delete message. Note
that, at least for now, we still have Modem share ownership of the
Cellular instance. If there is a potential race between Modem
destruction and the handling of a corresponding RTNL link delete
message, we would want to ensure that the behavior remains unchanged
to reduce the impact of this change.

A future change will rename DeviceInfo::RemoveInfo to
DeregisterDevice. This was not done in the change for the sake of
clarity.

BUG=chromium:1035061
TEST=All unit tests are passing.
TEST=On a nautilus: `rmmod cdc_mbim && rmmod cdc_wdm`
     See in logs: "Device destructed: wwan0"
TEST=Ensure suspend-resume behavior is unchanged for cellular.

Change-Id: If087d00de58be8030c26b1e251c45540fa67499a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2008092
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>