cryptohome: CheckKeyEx for ephemeral chall-resp
This adds the support of the CheckKeyEx operation for the
challenge-response mounts (a.k.a. smart card based login) in the
ephemeral mode. The implementation performs the "lightweight"
authentication that doesn't use the vault keyset or heavy TPM
operations.

As a side effect, this CL also significantly speeds up unlocking the
locked session for all smart card based users, even when the ephemeral
mode is not enabled.

BUG=chromium:1044558
TEST=enable ephemeral mode, log in via a smart card, lock session, verify that unlocking works

Change-Id: I35d6734de34f7f7b70762dfc9628a818f4127300
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2011106
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>