cros-disks: Prevent FUSE mount points from being unmounted twice
Previously, it was possible for FUSE mount points to be unmounted twice;
once by a browser request over D-Bus, and once by the process exit
watcher. This could lead to a situation where the mount point was
unmounted by browser, mounted again by the browser, and then the process
exit watcher from the first mount point is run. In this case, the exit
watcher will unmount the second mount by the browser instead of only
acting on the first. This would cause exFAT and NTFS filesystem to be
unavailable after waking from suspend.

BUG=chromium:995128,chromium:1026585,b:142372547
TEST=cros_run_unit_tests --board=chell --packages cros-disks
TEST=test_that <DUT_IP> 'e:platform_CrosDisks.*'
TEST=Insert NTFS formatted USB drive, lock screen,
     unplug and plug USB drive, unlock screen. Files on USB drive
     should be accessible in Files App.
TEST=Insert NTFS formatted USB drive, run 'suspend_stress_test -c 1000',
     Files on USB drive should be accessible in Files App after stress
     test.

Change-Id: I289f09877d09e73b37430710dab888cac1efb051
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2011483
Commit-Queue: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>