arc: networkd: Crostini service manages it's own devices.
This patch is the next in the series that lead to the removal of
DeviceManager. This change begins the process of having each service
manage its own devices starting with Crostini. The DM is still necessary
to provide so the service can register the default interface change
callback. Eventually this feature will probably just be widened on the
shill client itself.

BUG=b:145483351
TEST=units, manual on atlas

Change-Id: I7f8e7216828bda36167cf15c2790dd32a982acd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2012248
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>