crash: Make crash tests more efficient.
In crash tests, wait shorter on obtaining the global lock. This
was the bottleneck of the test exercising the global lock behavior.

This patch changes the timeout of attempting the lock to 1 second
in crash tests. Also changes the interval of lock attempts from
10 seconds to 1 second. The latter change affects the production
behavior as well, but it should not affect the performance because
attempting a file lock is a light-weight operation.

Also update the name of the crash test with the newer one.

BUG=chromium:978255
TEST=CQ

Change-Id: Iba5c297b41f7b438ccfe9da9d0db5d1294d5784c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2012243
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>