cryptohome: Implement lightweight chall-resp check
Implement the "VerifyKey" operation in the challenge-response component
of cryptohomed, which allows to perform the lightweight authentication
without the vault keyset and without heavy TPM operations.

This paves the way for implementing the lightweight shortcut for the
CheckKeyEx operation that makes the unlock working correctly for
ephemeral challenge-response users and working faster for all
challenge-response users.

BUG=chromium:1044558
TEST=existing tests

Change-Id: I11bbae21d40cfd9be51efaecb3907c4441480300
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2014903
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>