vm_tools: Pass ipv4 address of container in Cicerone service
As part of the port forwarding work, Chromium requires access to the
ipv4 address of the container (user does not need to specify it).

To achieve this, cicerone passes the ipv4 address of the container in
the signal ContainerStarted. The crostini manager will save this ipv4
address upon receiving the signal.

BUG=chromium:848127
TEST=ChromeOS builds and correct ipv4 address passed (experimented)

Change-Id: I3abad7fba98b3b5c3c87ecefec1ad2a8d0d83eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2014646
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Tested-by: Matthew Chen <matterchen@google.com>
Commit-Queue: Matthew Chen <matterchen@google.com>