arc: networkd: Make system calls explicit in runner.
This patch started out as just refactoring iptables to run as root in
order to work with the bug fixed in 1.62. It became obvious that it
would be better to have the runner wrapper expose every call in this
manner as well so there is (a) more context and (b) less clutter in the
datapath caller.

BUG=chromium:1044832
TEST=units, manual testing on grunt/atlas

Change-Id: I925e12ff2a623c801a776cba1fc77c43162a59fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2022402
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>