cryptohome: Add supplemental OOP mounter reporting.
The existing metrics for out-of-process mounts have the challenge that
the errors that we want to report result in Chrome not starting, so we
need to wait until subsequent successful executions of Chrome for the
metrics to be uploaded.

In order to provide a clearer picture of what's going on, add some
non-destructive crash reporting to make sure we're not missing errors
because of lack of metrics reporting.

BUG=chromium:1045012
TEST=Make mounting fail, see crash report in /var/spool/crash.
TEST=Use crash_sender --dev to upload the crash to the staging server.
TEST=go/crash-staging shows the report at
https://crash-staging.corp.google.com/browse?q=&stbtiq=product:ChromeOS&reportid=0fcdabb2a5da1422&index=0#0

Change-Id: Ifefdf36cdbd42fa857482c4368373140fce48901
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2017162
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>