crash: Remount mounts as MS_SLAVE.
We encountered a bug where because of default behaviour in Minijail,
crash_reporter was holding on to a mount that was unmounted in the
root mount namespace, which caused subsequent errors.

Fix that by remounting mounts in the crash_reporter mount namespace
as MS_SLAVE so that things unmounted in the root namespace also get
unmounted in the crash_reporter namespace.

BUG=b:147609789
TEST=Crash something on purpose, crash dump gets generated.

Change-Id: Ibf0cb3fc43e320a4a7a4981e77b29daa90be3570
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2019774
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>