dlcservice: Fix infinite Install() blocker
When a signal from update_engine indicates an |Operation::IDLE|,
dlcservice believes that that installation is complete only if it were
waiting for an installation. |DlcService| goes ahead and completes the
installation for |DlcManager| to finish, at which point failures to
mount the DLC(s) images resulted in a blocking state preventing
|DlcManager| to ever handle future installations because it remained
locked in a busy installing state.

Prior steps to |FinishInstall()| are handled with the periodic checker
that is posted, but the point at which |FinishInstall()| is called the
periodic checker no longer posted.

Also, the period between |InitInstall()| and |FinishInstall()| were
missing the |CancelInstall()| calls on failure as periodic checker would
not be scheduled.

BUG=none
TEST=# unittests + tast + DUT

Change-Id: I1964f41002dcef96ed63bebdfb7a14b117a08716
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2021325
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>