shill: manage supplicant process proxy with SupplicantManager
We can only support a single SupplicantProcessProxy instance, because
the underlying D-Bus object proxy only provides a single
SetNameOwnerChangedCallback(). If we have multiple such proxies, the
last one to instantiate will win the callback and the other will never
see service restart notifications (appear/vanish). This regressed way
back in https://crrev.com/c/292967.

To resolve this, create a singleton SupplicantManager (stashed in
Manager) to hold the SupplicantProcessProxy instance -- it's totally OK
for multiple devices to interact with the same process proxy.

While we're add it, introduce a Scoped listener object, so we don't have
to manually call Add and Remove methods for this sort of thing.

BUG=chromium:1029162
TEST=unit tests
TEST=bring up WiFi (e.g., scanning); plug in Ethernet device (e.g., USB
     dongle); `restart wpasupplicant`; see WiFi recover / scanning works
TEST=802.1X Ethernet tests (e.g., new tast `network.Ethernet8021X*`)
TEST=suite:wifi_matfunc

Change-Id: Ia7272ec6a5053e319e535677be67dfc68cdaa920
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2021201
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>