shill: rtnl_message: handle length underflow
RTA_NEXT() performs subtraction on the length argument, and it's
important that this subtraction can produce a negative result, so that
the next iteration of the loop can understand that we ran off the end
(i.e., in RTA_OK()).

Note that:
 * the callers for this are already using 'int' for length
 * rtattr uses 'unsigned short' (i.e., 16-bit) for length, so we should
   already be violating previous bounds checks if the *real* length
   can't actually fit in 'int'

I unwittingly introduced this error when refactoring the parsing code
here in https://crrev.com/c/1991965.

BUG=chromium:1045487, chromium:1045803
TEST=fuzzer

Change-Id: I2b116c45584d3c30b9ea2e741b1729d115b8970d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2023290
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>