vm_tools: crostini_client: allow requesting digest when exporting a VM
For automatic deployment of Plugin VMs we expect administrators to
create a policy with a link to a VM image and SHA256 digest of that
image, so that we can verify that downloaded image is not damaged or
altered mid-flight. This change extends "vmc export" command to
request said SHA256 digest when exporting a VM:

	vmc export -d <VmName> <ImageName>

The digest is written into a <ImageName>.sha256.txt file in the same
folder (usually "~/Downloads") where exported image will be residing.

BUG=b:148281196
TEST="vmc export -d TestVm testvm.zip", ensure that digest matches
output of "sha256sum ~/Downloads/testvm.zip"

Change-Id: I433bfef342a67ca1d8837ee055c0295776d0243a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2039938
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>