vm_tools: vm_concierge: support generating digest when exporting a VM
For automatic deployment of Plugin VMs we expect administrators to
create a policy with a link to a VM image and SHA256 digest of that
image, so that we can verify that downloaded image is not damaged or
altered mid-flight. This change extends the "export VM" command to
generate said SHA256 digest when requested.

The digest is written into the 2nd file descriptor attached to the DBus
message. We are using a file descriptor instead of returning digest in a
reply or signal because export operation takes long time and CLI may
elect not to wait for it to complete before returning control to the
user and having it in a file means that it can be examined later.

BUG=b:148281196
TEST="vmc export -d TestVm testvm.zip", ensure that digest matches
output of "sha256sum ~/Downloads/testvm.zip"

Change-Id: I80b74881758eba7214a6d61ba16362e7737a6210
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2039937
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>