diagnostics: Return a single BatteryInfoPtr
Return a single BatteryInfoPtr instead of a std::vector of them when
BatteryFetcher::FetchBatteryInfo() is called. This simplifies the code
since it's unlikely that a device will have more than one battery.

This also resolves an issue where devices without a battery would return
an empty vector and the CrosHealthdMojoService would try to access the
first element in that vector.

BUG=chromium:1048808
BUG=chromium:1049600
TEST=FEATURES="test" emerge-$BOARD diagnostics
TEST=telem --category=battery

Change-Id: Id3b9ea23f2302caf5392a9560e2c88ce2ab4f462
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2042512
Reviewed-by: Paul Moy <pmoy@chromium.org>
Tested-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Jesse Schettler <jschettler@chromium.org>